backend/requirements.txt
django==4.2.16  # Pin to the latest 4.2.x version as of now
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
psycopg2-binary==2.9.10
pillow==11.2.1
gunicorn==23.0.0
django-filter==25.1
django-cors-headers==4.7.0
python-dotenv
dj-database-url==2.1.0